% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add.columns}
\alias{add.columns}
\title{Add new columns to data frame with NA values.}
\usage{
add.columns(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{names of columns as either symbols or strings.}
}
\value{
Data frame \code{df} with columns added.
}
\description{
Add new columns to data frame with NA values.
}
