% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{aggregate.polys}
\alias{aggregate.polys}
\alias{aggregate.polys.list}
\alias{aggregate.polys.SpatialPolygonsDataFrame}
\title{Aggregate polygon geometry and data by data column values.}
\usage{
aggregate.polys(polys, by, data = NULL)

\method{aggregate.polys}{list}(polys, by, data = NULL)

\method{aggregate.polys}{SpatialPolygonsDataFrame}(polys, by, data = NULL)
}
\arguments{
\item{polys}{polygons as either a list as returned by
\code{\link{sp.to.polys}} or SpatialPolygonsDataFrames.}

\item{by}{names of data columns to aggregate by.}

\item{data}{aggregated data to join to result.}
}
\value{
Polygons in same format as given.
}
\description{
\code{aggregate.polys} does not aggregate data, only joins it.
The caller is expected to separately aggregate one's data, see examples.
}
\examples{
\dontrun{
data = fold(municipalities$data, .(province), area=sum(area))
provinces = aggregate.polys(municipalities, by="province", data=data)
}
}
\seealso{
\code{\link{union.polys}}
}
