% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ancfile}
\alias{ancfile}
\title{Find file or directory from among ancestors.}
\usage{
ancfile(name, max.depth = 12)
}
\arguments{
\item{name}{names of files or directories to look for.
Wildcards as defined in \code{\link{Sys.glob}} are allowed.}

\item{max.depth}{maximum depth to search upwards.}
}
\value{
Path to file or directory relative to working directory.
}
\description{
Find file or directory from among ancestors.
}
