% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{antijoin}
\alias{antijoin}
\title{Subset one data frame to match what is not found in another.}
\usage{
antijoin(x, y, by = intersect(names(x), names(y)))
}
\arguments{
\item{x}{data frame to subset.}

\item{y}{reference data frame.}

\item{by}{names of columns to match.}
}
\value{
Data frame \code{x} with a subset of rows that don't have
a \code{by}-wise match in \code{y}. Columns of \code{x} are not modified.
}
\description{
Subset one data frame to match what is not found in another.
}
