% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{apply.breaks}
\alias{apply.breaks}
\title{Classify variable by breaks.}
\usage{
apply.breaks(x, class, lower, upper)
}
\arguments{
\item{x}{a numeric vector.}

\item{class}{a vector of classes.}

\item{lower}{a numeric vector of lower bounds of classes.}

\item{upper}{a numeric vector of upper bounds of classes.}
}
\value{
A vector the same length of \code{x} with values from \code{class}.
Elements outside given breaks will have values \code{NA}.
}
\description{
Classify variable by breaks.
}
\seealso{
\code{\link{cut}}
}
