% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{arrange}
\alias{arrange}
\title{Order a data frame by its colums.}
\usage{
arrange(df, ...)
}
\arguments{
\item{df}{a data frame.}

\item{...}{expressions in the context of \code{df} to order by.}
}
\value{
A data frame.
}
\description{
Order a data frame by its colums.
}
