% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{attach.holes}
\alias{attach.holes}
\title{Connect holes to their outer polygons.}
\usage{
attach.holes(polys)
}
\arguments{
\item{polys}{a data frame of polygons.}
}
\value{
A data frame of polygons.
}
\description{
Connect holes to their outer polygons.
}
