% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox.to.poly}
\alias{bbox.to.poly}
\title{Convert bounding box coordinates to a polygon data frame.}
\usage{
bbox.to.poly(bbox)
}
\arguments{
\item{bbox}{a list with items "xmin", "xmax", "ymin" and "ymax".}
}
\value{
A polygon data frame.
}
\description{
Convert bounding box coordinates to a polygon data frame.
}
