% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{bbox.to.sp}
\alias{bbox.to.sp}
\title{Convert bounding box coordinates to a SpatialPolygonsDataFrame.}
\usage{
bbox.to.sp(bbox)
}
\arguments{
\item{bbox}{a list with items "xmin", "xmax", "ymin" and "ymax".}
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
Convert bounding box coordinates to a SpatialPolygonsDataFrame.
}
