% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bigfold}
\alias{bigfold}
\title{Split data frame and calculate group-wise summaries.}
\usage{
bigfold(df, by, ...)
}
\arguments{
\item{df}{data frame to summarize.}

\item{by}{columns to split by, either quoted or strings.}

\item{...}{one or more keyword arguments with function calls
of summaries to calculate, e.g. \code{xmin=min(x), xmax=max(x)}.}
}
\value{
A data frame with one row for each unique combination of \code{by}
and columns \code{by} and \code{...} in sorted order of \code{by}.
}
\description{
Split data frame and calculate group-wise summaries.
}
