% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind.lines}
\alias{bind.lines}
\alias{bind.lines.data.frame}
\alias{bind.lines.list}
\alias{bind.lines.SpatialLinesDataFrame}
\title{Combine sets of lines to form a single collection.}
\usage{
bind.lines(...)

\method{bind.lines}{data.frame}(...)

\method{bind.lines}{list}(...)

\method{bind.lines}{SpatialLinesDataFrame}(...)
}
\arguments{
\item{...}{sets of lines as either data frames, lists as
returned by \code{\link{sp.to.lines}} or SpatialLinesDataFrames.}
}
\value{
Lines in same format as given.
}
\description{
Combine sets of lines to form a single collection.
}
