% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bounding.box}
\alias{bounding.box}
\alias{bounding.box.data.frame}
\alias{bounding.box.SpatialLinesDataFrame}
\alias{bounding.box.SpatialPointsDataFrame}
\alias{bounding.box.SpatialPolygonsDataFrame}
\title{Coordinates of a bounding box.}
\usage{
bounding.box(shapes)

\method{bounding.box}{data.frame}(shapes)

\method{bounding.box}{SpatialLinesDataFrame}(shapes)

\method{bounding.box}{SpatialPointsDataFrame}(shapes)

\method{bounding.box}{SpatialPolygonsDataFrame}(shapes)
}
\arguments{
\item{shapes}{a data frame of shapes or an \pkg{sp} object.}
}
\value{
A list with items "xmin", "xmax", "ymin" and "ymax".
}
\description{
Coordinates of a bounding box.
}
