% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer.lines}
\alias{buffer.lines}
\alias{buffer.lines.data.frame}
\alias{buffer.lines.list}
\alias{buffer.lines.SpatialLinesDataFrame}
\title{Create polygon buffers around lines.}
\usage{
buffer.lines(lines, width, byid, quadsegs = 10, ...)

\method{buffer.lines}{data.frame}(lines, width, byid, quadsegs = 10, ...)

\method{buffer.lines}{list}(lines, width, byid, quadsegs = 10, ...)

\method{buffer.lines}{SpatialLinesDataFrame}(lines, width, byid,
  quadsegs = 10, ...)
}
\arguments{
\item{lines}{a data frame of lines or a list as returned by
\code{\link{sp.to.lines}} or a SpatialLinesDataFrame object.}

\item{width}{amount in coordinate units to offset by.}

\item{byid}{\code{TRUE} to produce separate objects.}

\item{quadsegs}{amount of segments to use in a quarter circle.}

\item{...}{passed to \code{\link[rgeos]{gBuffer}}.}
}
\value{
Polygons as either a list or
a SpatialPolygonsDataFrame object, depending on given input.
}
\description{
Create polygon buffers around lines.
}
