% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer.points}
\alias{buffer.points}
\alias{buffer.points.data.frame}
\alias{buffer.points.SpatialPointsDataFrame}
\title{Create polygon buffers around points.}
\usage{
buffer.points(points, width, byid, quadsegs = 10, ...)

\method{buffer.points}{data.frame}(points, width, byid, quadsegs = 10, ...)

\method{buffer.points}{SpatialPointsDataFrame}(points, width, byid, quadsegs = 10, ...)
}
\arguments{
\item{points}{a data frame of points or
a SpatialPointsDataFrame object.}

\item{width}{amount in coordinate units to offset by.}

\item{byid}{\code{TRUE} to produce separate objects.}

\item{quadsegs}{amount of segments to use in a quarter circle.}

\item{...}{passed to \code{\link[rgeos]{gBuffer}}.}
}
\value{
Polygons as either a list or
a SpatialPolygonsDataFrame object, depending on given input.
}
\description{
Create polygon buffers around points.
}
