% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer.polys}
\alias{buffer.polys}
\alias{buffer.polys.data.frame}
\alias{buffer.polys.list}
\alias{buffer.polys.SpatialPolygonsDataFrame}
\title{Create polygon buffers around polygons.}
\usage{
buffer.polys(polys, width, byid, quadsegs = 10, ...)

\method{buffer.polys}{data.frame}(polys, width, byid, quadsegs = 10, ...)

\method{buffer.polys}{list}(polys, width, byid, quadsegs = 10, ...)

\method{buffer.polys}{SpatialPolygonsDataFrame}(polys, width, byid,
  quadsegs = 10, ...)
}
\arguments{
\item{polys}{a data frame of polygon outlines or a list as returned by
\code{\link{sp.to.polys}} or a SpatialPolygonsDataFrame object.}

\item{width}{amount in coordinate units to offset by.}

\item{byid}{\code{TRUE} to produce separate objects.}

\item{quadsegs}{amount of segments to use in a quarter circle.}

\item{...}{passed to \code{\link[rgeos]{gBuffer}}.}
}
\value{
Polygons as either a list or
a SpatialPolygonsDataFrame object, depending on given input.
}
\description{
Create polygon buffers around polygons.
}
