% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{bumpTxtProgressBar}
\alias{bumpTxtProgressBar}
\title{Bump progress in progress bar.}
\usage{
bumpTxtProgressBar(bar, n = 1)
}
\arguments{
\item{bar}{progress bar object.}

\item{n}{amount to bump by.}
}
\description{
Bump progress in progress bar.
}
\seealso{
\code{\link{setTxtProgressBar}}
}
