% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{catn}
\alias{catn}
\title{Concatenate output separated by newlines.}
\usage{
catn(..., sep = "\\n")
}
\arguments{
\item{...}{passed to \code{\link{cat}}.}

\item{sep}{strings to append after each element.}
}
\description{
Concatenate output separated by newlines.
}
