% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cd}
\alias{cd}
\title{Set working directory and list files.}
\usage{
cd(...)
}
\arguments{
\item{...}{passed to \code{\link{setwd}}.}
}
\value{
See \code{\link{setwd}}.
}
\description{
Set working directory and list files.
}
