% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check.na}
\alias{check.na}
\title{Check for \code{NA} values and print result.}
\usage{
check.na(...)
}
\arguments{
\item{...}{data frames or objects coercible to data frames.}
}
\description{
Check for \code{NA} values and print result.
}
