% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chop}
\alias{chop}
\title{Split data frame to a list of data frames.}
\usage{
chop(x, id)
}
\arguments{
\item{x}{a data frame.}

\item{id}{a vector of values to chop by.}
}
\value{
A list of data frames.
}
\description{
Split data frame to a list of data frames.
}
