% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{circles}
\alias{circles}
\title{Construct polygon approximations of circles.}
\usage{
circles(xc, yc, radius, n = 100)
}
\arguments{
\item{xc}{x-coordinates of circle centroids.}

\item{yc}{y-coordinates of circle centroids.}

\item{radius}{radii of circles.}

\item{n}{amount of points in one circle polygon.}
}
\value{
A data frame of polygons.
}
\description{
Construct polygon approximations of circles.
}
