% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{classify}
\alias{classify}
\title{Form unique identifiers from combinations of values.}
\usage{
classify(...)
}
\arguments{
\item{...}{vectors of the same length.}
}
\value{
An integer-value vector the same length as arguments.
}
\description{
\code{classify} is what is also known as a "pairing function",
but not an invertible one. This also means that return values
for individual elements depend on other elements given as arguments
and the values of individual elements are thus not comparable
across calls.
}
