% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip.polys}
\alias{clip.polys}
\alias{clip.polys.data.frame}
\alias{clip.polys.list}
\alias{clip.polys.SpatialPolygonsDataFrame}
\title{Clip polygons to a bounding polygon.}
\usage{
clip.polys(polys, ...)

\method{clip.polys}{data.frame}(polys, ...)

\method{clip.polys}{list}(polys, ...)

\method{clip.polys}{SpatialPolygonsDataFrame}(polys, ...)
}
\arguments{
\item{polys}{a data frame of polygon outlines or a list as returned by
\code{\link{sp.to.polys}} or a SpatialPolygonsDataFrame object.}

\item{...}{bounds either as a data frame of polygon outlines,
a SpatialPolygonsDataFrame object or bounding box coordinates
\code{(xmin=..., xmax=..., ymin=..., ymax=...)}.}
}
\value{
Polygons in same format as \code{polys}.
}
\description{
Clip polygons to a bounding polygon.
}
