% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip.segments}
\alias{clip.segments}
\title{Clip segments to a bounding polygon.}
\usage{
clip.segments(segments, ...)
}
\arguments{
\item{segments}{a data frame of segments.}

\item{...}{bounds either as a data frame of polygon outlines,
a SpatialPolygonsDataFrame object or bounding box coordinates
\code{(xmin=..., xmax=..., ymin=..., ymax=...)}.}
}
\value{
A list of segments.
}
\description{
Clip segments to a bounding polygon.
}
