% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{colnames.prefix}
\alias{colnames.prefix}
\title{Add prefix to column names.}
\usage{
colnames.prefix(x, prefix)
}
\arguments{
\item{x}{a matrix-like object.}

\item{prefix}{a character string.}
}
\value{
A character vector.
}
\description{
Add prefix to column names.
}
