% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{colnames.suffix}
\alias{colnames.suffix}
\title{Add suffix to column names.}
\usage{
colnames.suffix(x, suffix)
}
\arguments{
\item{x}{a matrix-like object.}

\item{suffix}{a character string.}
}
\value{
A character vector.
}
\description{
Add suffix to column names.
}
