% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{connect.consecutive.lines}
\alias{connect.consecutive.lines}
\title{Connect consecutive line objects.}
\usage{
connect.consecutive.lines(lines, by, threshold = 0)
}
\arguments{
\item{lines}{a list as returned by \code{\link{sp.to.lines}}.}

\item{by}{names of columns in data that must match in order to join.}

\item{threshold}{maximum allowed distance of line-ends.}
}
\value{
A list of lines.
}
\description{
Connect consecutive line objects.
}
