% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{cores.available}
\alias{cores.available}
\title{Query amount of logical CPU cores designated for use.}
\usage{
cores.available()
}
\value{
Amount of cores.
}
\description{
Query amount of logical CPU cores designated for use.
}
