% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyr.R
\name{__ddply__}
\alias{__ddply__}
\alias{ddply}
\title{ddply from plyr.}
\usage{
ddply(
  .data,
  .variables,
  .fun = NULL,
  ...,
  .progress = "none",
  .inform = FALSE,
  .drop = TRUE,
  .parallel = FALSE,
  .paropts = NULL
)
}
\arguments{
\item{.data}{data frame to be processed}

\item{.variables}{variables to split data frame by, as as.quoted variables, a
formula or character vector}

\item{.fun}{function to apply to each piece}

\item{...}{See \code{\link[plyr]{ddply}} in \pkg{plyr}.}

\item{.progress}{name of the progress bar to use, see
\code{\link[plyr]{create_progress_bar}}.}

\item{.inform}{produce informative error messages?}

\item{.drop}{should combinations of variables that do not appear in the input
data be preserved (\code{FALSE}) or dropped (\code{TRUE}, default)}

\item{.parallel}{not applicable, see \code{\link{mcddply}} instead.}

\item{.paropts}{not applicable, see \code{\link{mcddply}} instead.}
}
\description{
ddply from plyr.

Split data frame and calculate group-wise summaries.
}
