% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians.}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{numeric object of degrees.}
}
\value{
A numeric object of radians.
}
\description{
Convert degrees to radians.
}
