% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dfsas}
\alias{dfsas}
\title{Create a data frame keeping strings as strings.}
\usage{
dfsas(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{passed to \code{\link{data.frame}}.}

\item{stringsAsFactors}{\code{TRUE} to convert strings to factors.}
}
\value{
A data frame.
}
\description{
Create a data frame keeping strings as strings.
}
\seealso{
\code{\link{data.frame}}
}
