% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{diff.polys}
\alias{diff.polys}
\alias{diff.polys.data.frame}
\alias{diff.polys.list}
\alias{diff.polys.SpatialPolygonsDataFrame}
\title{Remove areas of second polygons from first.}
\usage{
diff.polys(polys1, polys2)

\method{diff.polys}{data.frame}(polys1, polys2)

\method{diff.polys}{list}(polys1, polys2)

\method{diff.polys}{SpatialPolygonsDataFrame}(polys1, polys2)
}
\arguments{
\item{polys1}{Polygons to operate on.}

\item{polys2}{Polygons to remove from \code{polys1}.}
}
\value{
Polygons in same format as given.
}
\description{
Polygons can be given as either outline data frames, lists as
returned by \code{\link{sp.to.polys}} or SpatialPolygonsDataFrames,
but both \code{polys1} and \code{polys2} should be in the same format.
}
