% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{.add.standard.tiles}
\alias{.add.standard.tiles}
\title{Add standard tile sources and selector to map.}
\usage{
.add.standard.tiles(map)
}
\arguments{
\item{map}{A HTML widget object from \code{\link[leaflet]{leaflet}} function.}
}
\description{
Add standard tile sources and selector to map.
}
