% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.}
\alias{.}
\title{Quote variables to create a list of unevaluated expressions.}
\usage{
.(..., .env = parent.frame())
}
\arguments{
\item{...}{unevaluated expressions to be recorded.}

\item{.env}{environment in which to evaluate.}
}
\value{
List of symbol and language primitives.
}
\description{
Don't go crazy using \code{.} all over the place! \code{.} is ported over
from \pkg{plyr} for use in conjunction with \code{\link{ddply}}.
}
