% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{downclass}
\alias{downclass}
\title{Use minimal classes for data frame columns.}
\usage{
downclass(df)
}
\arguments{
\item{df}{a data frame.}
}
\value{
Data frame \code{df} with different column classes.
}
\description{
Use minimal classes for data frame columns.
}
