% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{download.file.retry}
\alias{download.file.retry}
\title{Download file, retrying in case of failure.}
\usage{
download.file.retry(url, destfile, max.tries = 3, interval = 5, ...)
}
\arguments{
\item{url}{URL of file to download.}

\item{destfile}{name where the downloaded file is saved.}

\item{max.tries}{maximum amount of tries.}

\item{interval}{interval in seconds to wait between tries.}

\item{...}{passed to \code{\link{download.file}}.}
}
\description{
Download file, retrying in case of failure.
}
