% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmap.R
\name{download.webmap}
\alias{download.webmap}
\title{Download and merge webmap tiles to form one image file.}
\usage{
download.webmap(
  template,
  zoom,
  xmin,
  xmax,
  ymin,
  ymax,
  fname,
  size = 256,
  epsg = 4326,
  wait.range = c(1, 3)
)
}
\arguments{
\item{template}{URL template with variables $X, $Y and $Z,
replaced with X-tile number, Y-tile number and zoom level,
e.g. \code{http://tile.openstreetmap.org/$Z/$X/$Y.png}.}

\item{zoom}{zoom level to download.}

\item{xmin}{WGS 84 coordinate of  west edge of area to download.}

\item{xmax}{WGS 84 coordinate of  east edge of area to download.}

\item{ymin}{WGS 84 coordinate of south edge of area to download.}

\item{ymax}{WGS 84 coordinate of north edge of area to download.}

\item{fname}{name of image file to save result to.}

\item{size}{pixel size of tiles (usually 256, 512 or 1024).}

\item{epsg}{projection code of written MapInfo file:
either 4326 for WGS 84 or 3857 for Web Mercator.}

\item{wait.range}{range of seconds from which to sample time to wait
between downloading successive 5x5 blocks of tiles. Using zeros
will skip waiting, but you will risk being banned by the tile server.}
}
\description{
Requires \code{montage} from \code{imagemagick}.
}
