% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop.points}
\alias{drop.points}
\alias{drop.points.data.frame}
\alias{drop.points.SpatialPointsDataFrame}
\title{Drop points outside bounding polygon.}
\usage{
drop.points(points, ...)

\method{drop.points}{data.frame}(points, ...)

\method{drop.points}{SpatialPointsDataFrame}(points, ...)
}
\arguments{
\item{points}{a data frame of points or a SpatialPointsDataFrame object.}

\item{...}{bounds either as a data frame of polygon outlines,
a SpatialPolygonsDataFrame object or bounding box coordinates
\code{(xmin=..., xmax=..., ymin=..., ymax=...)}.}
}
\value{
Points in same format as \code{points}.
}
\description{
Drop points outside bounding polygon.
}
