% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop.shapes}
\alias{drop.shapes}
\title{Drop shapes outside bounding box.}
\usage{
drop.shapes(shapes, xmin, xmax, ymin, ymax)
}
\arguments{
\item{shapes}{a data frame of shapes.}

\item{xmin}{west edge of the bounding box.}

\item{xmax}{east edge of the bounding box.}

\item{ymin}{south edge of the bounding box.}

\item{ymax}{north edge of the bounding box.}
}
\value{
A data frame of shapes.
}
\description{
Drop shapes outside bounding box.
}
