% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{eucd}
\alias{eucd}
\title{Euclidean distance.}
\usage{
eucd(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x-coordinates of first points.}

\item{y1}{y-coordinates of first points.}

\item{x2}{x-coordinates of second points.}

\item{y2}{y-coordinates of second points.}
}
\value{
A numeric object.
}
\description{
Euclidean distance.
}
\seealso{
\code{\link{eucd2}}
}
