% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{extend.bbox}
\alias{extend.bbox}
\title{Extend coordinates of a bounding box.}
\usage{
extend.bbox(bbox, buffer)
}
\arguments{
\item{bbox}{a list with items "xmin", "xmax", "ymin" and "ymax".}

\item{buffer}{amount to exceed by.}
}
\value{
A bounding box.
}
\description{
Extend coordinates of a bounding box.
}
