% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor2chr}
\alias{factor2chr}
\title{Convert factors to character without loss of information}
\usage{
factor2chr(x)
}
\arguments{
\item{x}{a vector containing factors.}
}
\value{
vector \code{x} converted to characters
}
\description{
Convert factors to character without loss of information
}
