% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor2int}
\alias{factor2int}
\title{Convert factors to integer without loss of information}
\usage{
factor2int(x)
}
\arguments{
\item{x}{a vector containing factors.}
}
\value{
vector \code{x} converted to integers
}
\description{
Convert factors to integer without loss of information
}
