% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor2num}
\alias{factor2num}
\title{Convert factors to numeric without loss of information}
\usage{
factor2num(x)
}
\arguments{
\item{x}{a vector containing factors.}
}
\value{
vector \code{x} converted to numerics.
}
\description{
Convert factors to numeric without loss of information
}
