% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{file_location}
\alias{file_location}
\title{Return directory of current file.}
\usage{
file_location()
}
\value{
A character string
}
\description{
Use only when called from Rscript or sourced via R console!
}
