% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{find.containers}
\alias{find.containers}
\title{Find polygons in which points are located.}
\usage{
find.containers(x, y, outlines)
}
\arguments{
\item{x}{a numeric vector of X-coordinates of points.}

\item{y}{a numeric vector of Y-coordinates of points.}

\item{outlines}{a data frame of polygon outlines.}
}
\value{
A data frame of "eid" and "sid" values for each point.
For points not inside any polygon, the value will be \code{NA}.
}
\description{
For the usual case of merging data from containing polygons,
\code{\link{points.in.polys}} should be more convenient.
\code{\link{find.containers}} is better if you need the full geometry,
i.e. both eid and sid values of outer polygons.
}
\seealso{
\code{\link{points.in.polys}}
}
