% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{flip.opposing.segments}
\alias{flip.opposing.segments}
\title{Reverse direction of opposing segments.}
\usage{
flip.opposing.segments(segments)
}
\arguments{
\item{segments}{a data frame of segments.}
}
\value{
Data frame \code{segments} with some segments flipped.
}
\description{
If there are two segments between a pair of nodes going in opposite
directions, flip either one of those so that they both go in the same
direction. The usual use case for this is prior to aggregation, where only
one value is desired between a pair of nodes.
}
