% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{flip_segments}
\alias{flip_segments}
\title{Flip direction of segments.}
\usage{
flip_segments(segments)
}
\arguments{
\item{segments}{a data frame of segments.}
}
\value{
a data frame of segments, with endpoints flipped.
}
\description{
Flip direction of segments.
}
