% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fname.clean}
\alias{fname.clean}
\title{Sanitize filenames.}
\usage{
fname.clean(fname)
}
\arguments{
\item{fname}{a character vector of filenames.}
}
\value{
A character vector.
}
\description{
Sanitize filenames.
}
