% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{fulljoin}
\alias{fulljoin}
\title{Join two data frames together.}
\usage{
fulljoin(x, y, by = intersect(names(x), names(y)), missing = NA)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{by}{names of columns to join by.}

\item{missing}{value to use in new columns for rows
found in \code{x}, but not in \code{y} or vice versa.}
}
\value{
A data frame with all \code{by}-wise
unique rows of both \code{x} and \code{y}.
}
\description{
Join two data frames together.
}
