% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extension from a filepath}
\usage{
get_file_ext(file_path)
}
\arguments{
\item{file_path}{a file path}
}
\value{
a character string
}
\description{
Get file extension from a filepath
}
