% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{grid.to.sp}
\alias{grid.to.sp}
\title{Convert a grid data frame to its \pkg{sp} equivalent.}
\usage{
grid.to.sp(grid, cell.size)
}
\arguments{
\item{grid}{a grid data frame.}

\item{cell.size}{size of grid cells.}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
Convert a grid data frame to its \pkg{sp} equivalent.
}
\seealso{
\code{\link[sp]{SpatialPolygonsDataFrame}}
}
