% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{hscale_segment}
\alias{hscale_segment}
\title{Horizontal scalebar for \pkg{ggplot2}.}
\usage{
hscale_segment(breaks, dy = 1/50, ...)
}
\arguments{
\item{breaks}{a data frame with columns "x" and "y".}

\item{dy}{tick height relative to entire width.}

\item{...}{passed to \code{\link[ggplot2]{geom_segment}}.}
}
\value{
Object from \code{\link[ggplot2]{geom_segment}}.
}
\description{
Horizontal scalebar for \pkg{ggplot2}.
}
