% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{hscale_text}
\alias{hscale_text}
\title{Horizontal scalebar text for \pkg{ggplot2}.}
\usage{
hscale_text(breaks, dy = 2/50, ...)
}
\arguments{
\item{breaks}{a data frame with columns "x", "y" and "label".}

\item{dy}{text offset relative to entire width.}

\item{...}{passed to \code{\link[ggplot2]{geom_text}}.}
}
\value{
Object from \code{\link[ggplot2]{geom_text}}.
}
\description{
Horizontal scalebar text for \pkg{ggplot2}.
}
