% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{index.shapes}
\alias{index.shapes}
\title{Determine shape ID values from data.}
\usage{
index.shapes(shapes)
}
\arguments{
\item{shapes}{a data frame of shapes.}
}
\value{
An integer vector of shape IDs.
}
\description{
Please note that elsewhere shape IDs may be indexed to be unique across one
entry, i.e. starting from one for each "eid", but \code{index.shapes} will
return shape IDs that are unique across all entries. This is done for
performance reasons and should not have any practical implications.
}
