% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet.R
\name{insert.monet}
\alias{insert.monet}
\title{Insert data into a MonetDB database table.}
\usage{
insert.monet(conn, table, value, fout = NULL, fin = NULL)
}
\arguments{
\item{conn}{An open database connection created with
\code{\link[DBI]{dbConnect}}.}

\item{table}{Name of table to insert into.}

\item{value}{A data frame to insert.}

\item{fout}{A path where to write a temporary CSV file if not the working
directory. The path must be writable by the script.}

\item{fin}{A path from where a temporary CSV file is read to the database if
not the working directory. The path must be readable by the database.}
}
\value{
\code{TRUE} if successful.
}
\description{
Insert data into a MonetDB database table.
}
