% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.odd}
\alias{is.odd}
\title{Test whether or not input is divisible by two.}
\usage{
is.odd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A logical vector.
}
\description{
Test whether or not input is divisible by two.
}
