% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{leftjoin}
\alias{leftjoin}
\title{Join two data frames together.}
\usage{
leftjoin(
  x,
  y,
  by = intersect(names(x), names(y)),
  overwrite = FALSE,
  missing = NA
)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{by}{names of columns to join by.}

\item{overwrite}{\code{TRUE} to replace existing columns in \code{x},
\code{FALSE} to only add new columns.}

\item{missing}{value to use in new columns for rows
found in \code{x}, but not in \code{y}.}
}
\value{
A data frame with as many rows as \code{x}.
}
\description{
Join two data frames together.
}
