% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{lineint}
\alias{lineint}
\title{Intersection points of two lines.}
\usage{
lineint(ix1, iy1, jx1, jy1, ix2, iy2, jx2, jy2)
}
\arguments{
\item{ix1}{X-coordinate of first point along line 1.}

\item{iy1}{Y-coordinate of first point along line 1.}

\item{jx1}{X-coordinate of second point along line 1.}

\item{jy1}{Y-coordinate of second point along line 1.}

\item{ix2}{X-coordinate of first point along line 2.}

\item{iy2}{Y-coordinate of first point along line 2.}

\item{jx2}{X-coordinate of second point along line 2.}

\item{jy2}{Y-coordinate of second point along line 2.}
}
\value{
A list of X and Y coordinates of intersection points.
}
\description{
Intersection points of two lines.
}
\references{
\url{http://mathworld.wolfram.com/Line-LineIntersection.html}
}
