% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{lines.to.sp}
\alias{lines.to.sp}
\title{Convert a line data frame to its \pkg{sp} equivalent.}
\usage{
\method{lines.to.sp}{not.an.s3.method}(lines, data = NULL, verbose = TRUE)
}
\arguments{
\item{lines}{a data frame of lines.}

\item{data}{a data frame of per-entry attributes.}

\item{verbose}{a logical indicating whether or not to print out progress.}
}
\value{
A \code{SpatialLinesDataFrame} object.
}
\description{
Convert a line data frame to its \pkg{sp} equivalent.
}
\seealso{
\code{\link[sp]{SpatialLinesDataFrame}}
}
