% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{list.emme.matrices}
\alias{list.emme.matrices}
\title{List matrices in Emme matrix file.}
\usage{
list.emme.matrices(fname)
}
\arguments{
\item{fname}{name of matrix file.}
}
\value{
A character vector of matrix names.
}
\description{
List matrices in Emme matrix file.
}
