% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load1.memo}
\alias{load1.memo}
\title{Reload one saved dataset keeping values cached.}
\usage{
load1.memo(fname)
}
\arguments{
\item{fname}{name of file to load. Use \code{NULL} to clear the cache.}
}
\value{
An object.
}
\description{
\code{load1.memo} is a memoized version of \code{\link{load1}}.
Objects loaded with \code{load1.memo} will stay in cache for
duration of the R session, unless size of the file changes
or cache is explicitly cleared. Be careful.
}
