% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load1}
\alias{load1}
\title{Reload one saved dataset.}
\usage{
load1(...)
}
\arguments{
\item{...}{passed to \code{\link{load}}.}
}
\value{
An object.
}
\description{
Reload a data set containing only one object and return that
object. Will fail if data set contains more than one object.
}
