% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{mcddply}
\alias{mcddply}
\title{Split data frame and calculate group-wise summaries.}
\usage{
mcddply(df, by, fun, ..., .progress = "text", .parallel = TRUE)
}
\arguments{
\item{df}{data frame to summarize.}

\item{by}{columns to split by, either quoted or strings.}

\item{fun}{function to apply to each piece.}

\item{...}{arguments passed to \code{fun}.}

\item{.progress}{"text" or "none".}

\item{.parallel}{\code{TRUE} to use all cores.}
}
\value{
A data frame.
}
\description{
\code{mcddply} is a parallel reimplementation of \code{\link{ddply}}.
Compared to \code{ddply} it is especially suitable for data frames
with a large amount of groups.
}
