% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{mclapply.stop}
\alias{mclapply.stop}
\title{Parallel version of lapply that stops after errors.}
\usage{
mclapply.stop(...)
}
\arguments{
\item{...}{passed to \code{\link{mclapply}}.}
}
\description{
Call \code{\link{mclapply}} and once all processes have completed,
check for possible errors and stop if errors found.
}
