% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mem.used}
\alias{mem.used}
\title{Return fraction of total memory in use.}
\usage{
mem.used()
}
\value{
Amount of total memory used, numeric between 0 and 1.
}
\description{
Return value is based on parsing output from the system command free.
In addition to free, awk is required. Return -1 in case of error.
Always return -1 on Windows.
}
