% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{messagef}
\alias{messagef}
\title{Generate a formatted diagnostic message.}
\usage{
messagef(...)
}
\arguments{
\item{...}{passed to \code{\link{sprintf}}.}
}
\description{
Generate a formatted diagnostic message.
}
