% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{messagef0}
\alias{messagef0}
\title{Generate a formatted diagnostic message without a newline.}
\usage{
messagef0(...)
}
\arguments{
\item{...}{passed to \code{\link{sprintf}}.}
}
\description{
Generate a formatted diagnostic message without a newline.
}
