% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mix.colors}
\alias{mix.colors}
\title{Average two color values.}
\usage{
mix.colors(x, y, pos = 0.5)
}
\arguments{
\item{x}{a color as understood by \code{\link{col2rgb}}.}

\item{y}{a color as understood by \code{\link{col2rgb}}.}

\item{pos}{interpolation position between \code{x} and \code{y}.}
}
\value{
A color.
}
\description{
Average two color values.
}
