% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Test against inclusion, opposite of \code{\%in\%}.}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched}

\item{table}{vector or \code{NULL}: the values to be matched against}
}
\value{
A logical vector.
}
\description{
Test against inclusion, opposite of \code{\%in\%}.
}
