% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{non.na}
\alias{non.na}
\title{Discard \code{NA} elements.}
\usage{
non.na(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
\code{x} with \code{NA} elements discarded.
}
\description{
Discard \code{NA} elements.
}
