% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers.iqr}
\alias{outliers.iqr}
\title{Find outliers based on inter-quartile range.}
\usage{
outliers.iqr(x, lower = TRUE, upper = TRUE, coeff = 1.5,
  na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{lower}{\code{TRUE} to check lower end for outliers.}

\item{upper}{\code{TRUE} to check upper end for outliers.}

\item{coeff}{maximum IQR  allowed for non-outlier elements.}

\item{na.rm}{\code{TRUE} to ignore \code{NA} elements.}
}
\value{
A logical vector indicating which of \code{x} are outliers.
}
\description{
Find outliers based on inter-quartile range.
}
