% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad}
\alias{pad}
\title{Pad string.}
\usage{
pad(x, pad = " ", n)
}
\arguments{
\item{x}{a vector coercible to character.}

\item{pad}{a padding string.}

\item{n}{length of padded string.}
}
\value{
A character vector.
}
\description{
Pad string.
}
