% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pasten}
\alias{pasten}
\title{Join string by a newline separator.}
\usage{
pasten(..., sep = "\\n")
}
\arguments{
\item{...}{passed on to \code{\link{paste}}.}

\item{sep}{separator string.}
}
\value{
A character vector.
}
\description{
Join string by a newline separator.
}
