% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{perp.angle}
\alias{perp.angle}
\title{Angles perpendicular to a segments.}
\usage{
perp.angle(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x-coordinates of first points.}

\item{y1}{y-coordinates of first points.}

\item{x2}{x-coordinates of second points.}

\item{y2}{y-coordinates of second points.}
}
\value{
A numeric object of angles in radians.
}
\description{
Angles perpendicular to a segments.
}
