% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{point.segment.dist}
\alias{point.segment.dist}
\title{Distances of points from segments.}
\usage{
point.segment.dist(p, a, b)
}
\arguments{
\item{p}{coordinates of points.}

\item{a}{coordinates of the start points of segments.}

\item{b}{coordinates of the end points of segments.}
}
\value{
A numeric vector of distances.
}
\description{
Either \code{p} or \code{a} and \code{b} can be vectorized, not both.
Coordinates should be given as either a numeric vector of length two
or a two-column matrix of X- and Y-coordinates.
}
\references{
\url{http://paulbourke.net/geometry/pointlineplane/}
}
