% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{points.in.polys}
\alias{points.in.polys}
\title{Find polygons in which points are located.}
\usage{
\method{points.in.polys}{not.an.s3.method}(x, y, outlines, data = NULL)
}
\arguments{
\item{x}{a numeric vector of X-coordinates of points.}

\item{y}{a numeric vector of Y-coordinates of points.}

\item{outlines}{a data frame of polygon outlines.}

\item{data}{a data frame of polygon data associated with outlines.}
}
\value{
If \code{data} given, an integer value vector of indices of
points in \code{data}. For points not in any polygon, value will be
\code{NA}. If data not given, a logical vector.
}
\description{
Find polygons in which points are located.
}
\seealso{
\code{\link{find.containers}}
}
