% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{points.to.sp}
\alias{points.to.sp}
\title{Convert a point data frame to its \pkg{sp} equivalent.}
\usage{
\method{points.to.sp}{not.an.s3.method}(points)
}
\arguments{
\item{points}{a data frame of points.}
}
\value{
A \code{SpatialPointsDataFrame} object.
}
\description{
Convert a point data frame to its \pkg{sp} equivalent.
}
\seealso{
\code{\link[sp]{SpatialPointsDataFrame}}
}
