% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{points_to_lines}
\alias{points_to_lines}
\title{Convert a data frame of groups of points into lines}
\usage{
points_to_lines(points, group_id)
}
\arguments{
\item{points}{a data frame of points with x and y coordinate.}

\item{group_id}{column identifying the group of lines to be formed in one
line.}
}
\value{
A list with two data frames: data and lines.
}
\description{
Convert a data frame of groups of points into lines
}
\examples{
\dontrun{points.in.lines(points, trip_id)}
}
