% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{polys.to.sp}
\alias{polys.to.sp}
\title{Convert a polygon data frame to its \pkg{sp} equivalent.}
\usage{
polys.to.sp(outlines, data = NULL)
}
\arguments{
\item{outlines}{a data frame of polygon outlines.}

\item{data}{a data frame of per-entry attributes.}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
Convert a polygon data frame to its \pkg{sp} equivalent.
}
\seealso{
\code{\link[sp]{SpatialPolygonsDataFrame}}
}
