% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysimp.R
\name{polysimp.lines}
\alias{polysimp.lines}
\title{Simplify lines by dropping visually least significant points.}
\usage{
polysimp.lines(lines, epsilon, static = NULL, max.length = Inf,
  keep.double = TRUE, method = dp)
}
\arguments{
\item{lines}{a data frame of lines.}

\item{epsilon}{threshold, below which nodes will be removed.}

\item{static}{a logical vector indicating nodes which should not be removed.}

\item{max.length}{maximum length of line segments to form.}

\item{keep.double}{keep all duplicated nodes intact. This is necessary
to keep noded intersections in place, e.g. with a road network to maintain
connectivity for routing use.}

\item{method}{low-level polygon simplification function to use:
\code{\link{dp}} for Douglas-Peucker's algorithm or
\code{\link{visvalingam}} for Visvalingam's algorithm.}
}
\value{
A data frame of lines.
}
\description{
All duplicated nodes are kept intact.
}
