% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysimp.R
\name{polysimp.polys}
\alias{polysimp.polys}
\title{Simplify polygons by dropping visually least significant points.}
\usage{
polysimp.polys(polys, epsilon, static = NULL, max.length = Inf,
  keep.triple = TRUE, method = dp, pass = 1)
}
\arguments{
\item{polys}{a data frame of polygons.}

\item{epsilon}{threshold, below which nodes will be removed.}

\item{static}{a logical vector indicating nodes which should not be removed.}

\item{max.length}{maximum length of line segments to form.}

\item{keep.triple}{keep all triplicated nodes intact. For adjacent polygons
(e.g. administrative boundaries) this is necessary (in addition to the
two-pass process) to keep the polygons adjacent.}

\item{method}{low-level polygon simplification function to use:
\code{\link{dp}} for Douglas-Peucker's algorithm or
\code{\link{visvalingam}} for Visvalingam's algorithm.}

\item{pass}{Only true wizards should even consider setting this argument.}
}
\value{
A data frame of polygons.
}
\description{
Simplification is done in two passes. All nodes kept by any polygon
in the first pass are kept intact.
}
