% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysimp.R
\name{polysimp}
\alias{polysimp}
\title{Simplify polyline.}
\usage{
polysimp(coords, epsilon, static = NULL, max.length = Inf)
}
\arguments{
\item{coords}{a two-column matrix of coordinates of polyline nodes.}

\item{epsilon}{threshold, below which nodes will be removed.}

\item{static}{a logical vector indicating nodes which should not be removed.}

\item{max.length}{maximum length of line segments to form.}
}
\value{
A logical vector of rows to keep.
}
\description{
Simplify polyline.
}
