% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{post.gc}
\alias{post.gc}
\title{Collect garbage after another call.}
\usage{
post.gc(x)
}
\arguments{
\item{x}{returned after calling \code{\link{gc}}.}
}
\value{
\code{x} given.
}
\description{
Use as e.g. post.gc(rm(big.object)).
}
