% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress.due}
\alias{progress.due}
\title{Query permission to print progress output.}
\usage{
progress.due(interval, parallel = FALSE, n = NULL)
}
\arguments{
\item{interval}{amount of seconds between which to print.}

\item{parallel}{\code{TRUE} if task is being run in parallel processes.}

\item{n}{amount of processes used. Give only if different
from what is returned by \code{\link{cores.available}}.}
}
\value{
A logical value.
}
\description{
Query permission to print progress output.
}
