% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress.eta}
\alias{progress.eta}
\title{Print progress of running task.}
\usage{
progress.eta(time.start, n, total, suffix = "")
}
\arguments{
\item{time.start}{start time as returned by \code{\link{Sys.time}}.}

\item{n}{current item.}

\item{total}{total amount of items.}

\item{suffix}{additional text to include.}
}
\description{
Print progress of running task.
}
