% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress.final}
\alias{progress.final}
\title{Print statistics of completed task.}
\usage{
progress.final(time.start)
}
\arguments{
\item{time.start}{start time as returned by \code{\link{Sys.time}}.}
}
\description{
Print statistics of completed task.
}
