% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj.mapinfo}
\alias{proj.mapinfo}
\title{MapInfo-strings of supported projections.}
\usage{
\method{proj.mapinfo}{not.an.s3.method}(epsg)
}
\arguments{
\item{epsg}{projection EPSG code.}
}
\value{
A string or \code{NA}.
}
\description{
MapInfo-strings of supported projections.
}
