% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj.name}
\alias{proj.name}
\title{Names of supported projections.}
\usage{
\method{proj.name}{not.an.s3.method}(epsg)
}
\arguments{
\item{epsg}{projection EPSG code.}
}
\value{
A string or \code{NA}.
}
\description{
Names of supported projections.
}
