% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj.ogr.mapinfo}
\alias{proj.ogr.mapinfo}
\title{Projection strings to write into MapInfo files using ogr2ogr.}
\usage{
\method{proj.ogr.mapinfo}{not.an.s3.method}(epsg)
}
\arguments{
\item{epsg}{projection EPSG code.}
}
\value{
A string or \code{NA}.
}
\description{
Projection strings to write into MapInfo files using ogr2ogr.
}
