% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj.proj4}
\alias{proj.proj4}
\title{Proj4-strings of supported projections.}
\usage{
\method{proj.proj4}{not.an.s3.method}(epsg)
}
\arguments{
\item{epsg}{projection EPSG code.}
}
\value{
A string or \code{NA}.
}
\description{
Proj4-strings of supported projections.
}
