% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj.shape}
\alias{proj.shape}
\title{ESRI shape WKT-strings of supported projections.}
\usage{
\method{proj.shape}{not.an.s3.method}(epsg)
}
\arguments{
\item{epsg}{projection EPSG code.}
}
\value{
A string or \code{NA}.
}
\description{
ESRI shape WKT-strings of supported projections.
}
