% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees.}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{numeric object of radians.}
}
\value{
A numeric object of degrees.
}
\description{
Convert radians to degrees.
}
