% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{range.labels}
\alias{range.labels}
\title{Build numeric range labels for a plot legend.}
\usage{
\method{range.labels}{not.an.s3.method}(a, b, sep = "-")
}
\arguments{
\item{a}{a numeric or character vector of lower bounds of ranges.}

\item{b}{a numeric or character vector of upper bounds of ranges.}

\item{sep}{separator between lower and upper bounds.}
}
\value{
A quoted expression (or whatever). Just pass it as a
labels argument to a \pkg{ggplot2} scale function.
}
\description{
Build numeric range labels for a plot legend.
}
