% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rbind.via.file}
\alias{rbind.via.file}
\title{Combine data frames by row.}
\usage{
\method{rbind}{via.file}(...)
}
\arguments{
\item{...}{data frames.}
}
\value{
A data frame.
}
\description{
Combine data frames by row.
}
