% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{rbind_all}
\alias{rbind_all}
\title{Efficiently rbind multiple data frames.}
\usage{
rbind_all(dfs)
}
\arguments{
\item{dfs}{a list of data frames to combine.}
}
\value{
A data frame.
}
\description{
Efficiently rbind multiple data frames.
}
