% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{rbind_list}
\alias{rbind_list}
\title{Efficiently rbind multiple data frames.}
\usage{
rbind_list(...)
}
\arguments{
\item{...}{individual data frames to combine.}
}
\value{
A data frame.
}
\description{
Efficiently rbind multiple data frames.
}
