% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.any}
\alias{read.any}
\title{Read tabular data from files of various types.}
\usage{
read.any(fname, ...)
}
\arguments{
\item{fname}{name of file to read.}

\item{...}{passed to filetype-specific reading function.}
}
\value{
A data frame.
}
\description{
\code{read.any} guesses the filetype based on given filename's extension
and uses the appropriate function to read the given file. \code{read.any}
is suitable for cases where you have a vector of filenames, which you know
to contain standard form data, but where the format may vary.
}
