% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.delims}
\alias{read.delims}
\title{Read a file in delimited format keeping strings as strings.}
\usage{
read.delims(...)
}
\arguments{
\item{...}{passed to \code{\link{read.delim}}.}
}
\value{
A data frame.
}
\description{
Read a file in delimited format keeping strings as strings.
}
