% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{read.emme.matrices}
\alias{read.emme.matrices}
\title{Read matrices from Emme matrix file.}
\usage{
read.emme.matrices(fname)
}
\arguments{
\item{fname}{name of matrix file.}
}
\value{
A list of data frames.
}
\description{
Read matrices from Emme matrix file.
}
