% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{read.emme.matrix}
\alias{read.emme.matrix}
\title{Read matrix from Emme matrix file.}
\usage{
read.emme.matrix(fname, matrix)
}
\arguments{
\item{fname}{name of matrix file.}

\item{matrix}{name of matrix to read, e.g. "mf123".}
}
\value{
A data frame with columns "p", "q" and "value".
}
\description{
Read matrix from Emme matrix file.
}
