% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{read.emme.submatrix}
\alias{read.emme.submatrix}
\title{Read submatrix from emme output file.}
\usage{
read.emme.submatrix(fname)
}
\arguments{
\item{fname}{name of matrix file.}
}
\value{
A data frame.
}
\description{
Read submatrix from emme output file.
}
