% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg.R
\name{read.gpkg}
\alias{read.gpkg}
\title{This function has been removed from the package. Use `sf::st_read()` instead.}
\usage{
read.gpkg(fname, stringsAsFactors = FALSE, encoding = "UTF-8", ...)
}
\arguments{
\item{fname}{name of GeoPackage file}

\item{stringsAsFactors}{should character vector be converted to factors?}

\item{encoding}{character encoding used in code{fname}.}

\item{...}{passed to \code{\link[rgdal]{readOGR}}.}
}
\value{
A \code{Spatial*DataFrame} object.
}
\description{
This function has been removed from the package. Use `sf::st_read()` instead.
}
