% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{read.mapinfo.bbox}
\alias{read.mapinfo.bbox}
\title{Read bounding box from MapInfo TAB file.}
\usage{
read.mapinfo.bbox(fname)
}
\arguments{
\item{fname}{name of MapInfo TAB file.}
}
\value{
A list with items "xmin", "xmax", "ymin" and "ymax".
}
\description{
Read bounding box from MapInfo TAB file.
}
