% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{read.mapinfo.png}
\alias{read.mapinfo.png}
\title{Read and crop geographically referenced PNG-file saved by MapInfo.}
\usage{
read.mapinfo.png(fname, xmin = NULL, xmax = NULL, ymin = NULL,
  ymax = NULL, cache = FALSE)
}
\arguments{
\item{fname}{name of PNG-file.}

\item{xmin}{minimum of X-coordinates to crop image to.}

\item{xmax}{maximum of X-coordinates to crop image to.}

\item{ymin}{minimum of Y-coordinates to crop image to.}

\item{ymax}{maximum of Y-coordinates to crop image to.}

\item{cache}{\code{TRUE} to cache the uncropped PNG between calls.}
}
\value{
An image array.
}
\description{
Read and crop geographically referenced PNG-file saved by MapInfo.
}
