% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.matrix}
\alias{read.matrix}
\title{Read matrix from file.}
\usage{
read.matrix(fname, header = FALSE, sep = "\\t", ...)
}
\arguments{
\item{fname}{The name of the file to read.}

\item{header}{A logical value indicating whether the file contains the names
of the variables as its first line.}

\item{sep}{The field separator character. Values on each line of the file are
separated by this character.}

\item{...}{Further arguments to be passed to \code{\link[utils]{read.table}}.}
}
\value{
A matrix object.
}
\description{
See \code{\link[utils]{read.table}} for arguments.
}
