% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{read.mdb}
\alias{read.mdb}
\title{Read from MS Access database.}
\usage{
read.mdb(fname, table = NULL, query = NULL, ...)
}
\arguments{
\item{fname}{name of database file.}

\item{table}{name of table.}

\item{query}{SQL-query to use to fetch data. If not given, all
data in table is read using \code{"SELECT * FROM table"}.}

\item{...}{passed to \code{\link[RODBC]{sqlQuery}}.}
}
\value{
A data frame.
}
\description{
Uses ODBC, will fail on non-Windows systems.
}
