% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{read.shape}
\alias{read.shape}
\title{Read points, lines or polygons from a shapefile.}
\usage{
read.shape(fname, stringsAsFactors = FALSE, encoding = "Latin1", ...)
}
\arguments{
\item{fname}{name of shapefile.}

\item{stringsAsFactors}{should character vector be converted to factors?}

\item{encoding}{character encoding used in code{fname}.
Defaults to "Latin1", which is the original standard DBF encoding,
but in reality this will vary.}

\item{...}{passed to \code{\link[rgdal]{readOGR}}.}
}
\value{
A \code{Spatial*DataFrame} object.
}
\description{
Read points, lines or polygons from a shapefile.
}
