% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{read_geospatial_file}
\alias{read_geospatial_file}
\title{Read points, lines or polygons from a geospatial file}
\usage{
read_geospatial_file(fname)
}
\arguments{
\item{fname}{name of geospatial file (.shp, .gpkg or .TAB)}
}
\value{
A \code{Spatial*DataFrame} object.
}
\description{
Read points, lines or polygons from a geospatial file
}
