% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{recode}
\alias{recode}
\alias{recode.character}
\alias{recode.data.frame}
\title{Convert character encoding.}
\usage{
recode(x, from = NULL, to, ...)

\method{recode}{character}(x, from = NULL, to, ...)

\method{recode}{data.frame}(x, from = NULL, to, ...)
}
\arguments{
\item{x}{a character vector or a data frame.}

\item{from}{character encoding to convert from.
Use "default" as shorthand for \code{getOption("encoding")}
Can be \code{NULL} to try \code{\link{Encoding}}
(in case that fails, no conversion is done).}

\item{to}{character encoding to convert to.
Use "default" as shorthand for \code{getOption("encoding")}}

\item{...}{not in use}
}
\value{
A object of same type as given.
}
\description{
Convert character encoding.
}
