% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rename}
\alias{rename}
\title{Rename columns in data frame.}
\usage{
rename(df, ...)
}
\arguments{
\item{df}{a data frame.}

\item{...}{named argument pairs of form \code{from=to}, e.g.
\code{rename(df, x=y)} to rename \code{x} to \code{y} in \code{df}.
Strings and function calls are allowed on the right side, e.g.
\code{rename(df, x="y")} or \code{rename(df, x=sprintf(...))}.}
}
\value{
A data frame.
}
\description{
Rename columns in data frame.
}
