% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{reproject}
\alias{reproject}
\alias{reproject.data.frame}
\alias{reproject.matrix}
\alias{reproject.SpatialLinesDataFrame}
\alias{reproject.SpatialPointsDataFrame}
\alias{reproject.SpatialPolygonsDataFrame}
\title{Transform and convert between projections.}
\usage{
reproject(coords, from, to, names = c("x", "y"), fun = identity)

\method{reproject}{data.frame}(coords, from, to, names = c("x", "y"),
  fun = identity)

\method{reproject}{matrix}(coords, from, to, names = NULL,
  fun = identity)

\method{reproject}{SpatialLinesDataFrame}(coords, from, to, names = NULL,
  fun = NULL)

\method{reproject}{SpatialPointsDataFrame}(coords, from, to,
  names = NULL, fun = NULL)

\method{reproject}{SpatialPolygonsDataFrame}(coords, from, to,
  names = NULL, fun = NULL)
}
\arguments{
\item{coords}{a data frame with columns "x" and "y" or
a two-column matrix of X- and Y-coordinates or an \pkg{sp} object.}

\item{from}{EPSG code of current projection.}

\item{to}{EPSG code of desired projection.}

\item{names}{names of columns holding coordinates.
(Only relevant if \code{coords} is a data frame.)}

\item{fun}{function to apply to resulting coordinates (e.g. round).
(Only used if \code{coords} is a data.frame or a matrix.)}
}
\value{
A data frame or two-column matrix or an \pkg{sp} object.
}
\description{
Transform and convert between projections.
}
