% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{resample}
\alias{resample}
\title{Draw a random sample.}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{a vector of elements from which to choose.}

\item{...}{passed to \code{\link{sample}}.}
}
\value{
A vector of random elements.
}
\description{
Act as a better substitute for \code{\link{sample}}, which behaves
in a silly manner when the length of \code{x} is one.
}
