% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rm.gc}
\alias{rm.gc}
\title{Remove object from caller's environment and collect garbage.}
\usage{
rm.gc(..., list = character(0))
}
\arguments{
\item{...}{the objects to be removed, as names (unquoted)
or character strings (quoted).}

\item{list}{a character vector naming objects to be removed.}
}
\description{
Remove object from caller's environment and collect garbage.
}
