% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{round.grid}
\alias{round.grid}
\title{Round coordinate to grid cell center point.}
\usage{
\method{round.grid}{not.an.s3.method}(x, cell.size)
}
\arguments{
\item{x}{a numeric object.}

\item{cell.size}{size of cells in grid.}
}
\value{
A numeric or integer object.
}
\description{
Round coordinate to grid cell center point.
}
