% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{round.stoc}
\alias{round.stoc}
\title{Round stochastically with floor and ceiling.}
\usage{
\method{round.stoc}{not.an.s3.method}(x)
}
\arguments{
\item{x}{a numeric object.}
}
\value{
An integer object.
}
\description{
Round stochastically with floor and ceiling.
}
