% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{round.total}
\alias{round.total}
\title{Round stochastically with floor and ceiling, keeping total intact.}
\usage{
\method{round.total}{not.an.s3.method}(x)
}
\arguments{
\item{x}{a numeric object.}
}
\value{
An integer object.
}
\description{
Round stochastically with floor and ceiling, keeping total intact.
}
