% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number.}
\usage{
round_any(x, accuracy, fun = round)
}
\arguments{
\item{x}{a numeric vector.}

\item{accuracy}{number to round to.}

\item{fun}{rounding function: \code{\link{floor}}, \code{\link{ceiling}}
or \code{\link{round}}.}
}
\value{
A numeric or integer vector.
}
\description{
Round to multiple of any number.
}
