% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rows.along}
\alias{rows.along}
\title{Sequence of rows.}
\usage{
rows.along(x)
}
\arguments{
\item{x}{a matrix-like object.}
}
\value{
An integer vector.
}
\description{
Primary intended use case of \code{rows.along} is in
\code{for}-loops over the rows of a matrix or a data frame, where
the amount of rows is not known beforehand and might be zero
and thus 1:nrow(x) would be dangerous.
}
