% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{save.plot.pdf}
\alias{save.plot.pdf}
\title{Save a \pkg{ggplot2} plot as PDF.}
\usage{
save.plot.pdf(plot, fname, device = grDevices::cairo_pdf, ...)
}
\arguments{
\item{plot}{a \pkg{ggplot2} plot object.}

\item{fname}{name of PDF-file to produce.}

\item{device}{device to use.}

\item{...}{passed to \code{\link[ggplot2]{ggsave}}.}

\item{family}{name of font family to use.}
}
\description{
Save a \pkg{ggplot2} plot as PDF.
}
