% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{save.plot}
\alias{save.plot}
\title{Save a \pkg{ggplot2} plot.}
\usage{
save.plot(plot, fname, ...)
}
\arguments{
\item{plot}{a \pkg{ggplot2} plot object.}

\item{fname}{name of file to produce.}

\item{...}{passed to \code{\link[ggplot2]{ggsave}}.}
}
\description{
Save a \pkg{ggplot2} plot.
}
