% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale.segments}
\alias{scale.segments}
\title{Construct scaled polygons from segments.}
\usage{
\method{scale.segments}{not.an.s3.method}(
  segments,
  value,
  max.width,
  max.at = max(value, na.rm = TRUE),
  tol = 0,
  fun = identity
)
}
\arguments{
\item{segments}{a data frame of segments.}

\item{value}{a vector of values to scale widths by.}

\item{max.width}{maximum segment width in coordinate units.}

\item{max.at}{value at which width is scaled to \code{max.width}.}

\item{tol}{segment endpoint moving tolerance relative to length.}

\item{fun}{function to apply to coordinates (e.g. \code{\link{round}}).}
}
\value{
A data frame of polygons.
}
\description{
Endpoints of consecutive segments are moved to their intersection
if reasonably applicable to achieve a smoothly continuous form.
}
