% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{segment.label.angle}
\alias{segment.label.angle}
\title{Angle for segment labels.}
\usage{
segment.label.angle(segments)
}
\arguments{
\item{segments}{a data frame of segments.}
}
\value{
A vector of angles, in degrees.
}
\description{
Angle for segment labels.
}
