% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{segment.label.points}
\alias{segment.label.points}
\title{Label points for scaled segments.}
\usage{
segment.label.points(polys, vjust.above = -0.1, vjust.below = 1.1)
}
\arguments{
\item{polys}{a data frame of segment polygons as returned
by \code{\link{scale.segments}}.}

\item{vjust.above}{vjust specification for labels above, logically <= 0.}

\item{vjust.below}{vjust specification for labels below, logically >= 1.}
}
\value{
A data frame of points.
}
\description{
Label points for scaled segments.
}
