% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{segment.labels}
\alias{segment.labels}
\title{Labels for scaled segments.}
\usage{
segment.labels(segments, polys, vjust.above = -0.1, vjust.below = 1.1)
}
\arguments{
\item{segments}{a data frame of segments.}

\item{polys}{a data frame of segment polygons as returned
by \code{\link{scale.segments}}.}

\item{vjust.above}{vjust specification for labels above, logically <= 0.}

\item{vjust.below}{vjust specification for labels below, logically >= 1.}
}
\value{
A data frame of labels.
}
\description{
Labels for scaled segments.
}
