% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{segment.legend}
\alias{segment.legend}
\title{Polygon staircase legend for scaled segments.}
\usage{
segment.legend(x0, y0, values, width, height, xoffset)
}
\arguments{
\item{x0}{X-coordinate of point to start legend from.}

\item{y0}{Y-coordinate of point to start legend from.}

\item{values}{a vector of values at breaks. If sorted ascending, legend will
also ascend from \code{(x0,y0)}, if desceding, legend will also descend.
To acquire these from data, see e.g. \code{\link{pretty}}.}

\item{width}{width of the widest segment. If using
\code{\link{scale.segments}}, set this to the same as \code{max.width}
passed there. If \code{width} is positive, construct legend east of
\code{x0}, if negative, west of \code{x0}.}

\item{height}{height of one segment. If height is positive, construct
legend north of \code{y0}, if negative, south of \code{y0}.}

\item{xoffset}{amount of space to leave between polygon and associated
labels. This would depend on the amount of characters in \code{values}
and font size and horizontal justification used when plotting.}
}
\value{
A list of two data frames: "polys" and "labels".
}
\description{
Polygon staircase legend for scaled segments.
}
