% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{segments.to.lines}
\alias{segments.to.lines}
\title{Convert segments to lines}
\usage{
segments.to.lines(segments)
}
\arguments{
\item{segments}{a data frame of segments.}
}
\value{
A data frame of lines.
}
\description{
Convert segments to lines
}
