% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{segments.to.sp}
\alias{segments.to.sp}
\title{Convert a segment data frame to its \pkg{sp} equivalent.}
\usage{
segments.to.sp(segments)
}
\arguments{
\item{segments}{a data frame of segments.}
}
\value{
A \code{SpatialLinesDataFrame} object.
}
\description{
Convert a segment data frame to its \pkg{sp} equivalent.
}
\seealso{
\code{\link[sp]{SpatialLinesDataFrame}}
}
