% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{semijoin}
\alias{semijoin}
\title{Subset one data frame to match what is found in another.}
\usage{
semijoin(x, y, by = intersect(names(x), names(y)))
}
\arguments{
\item{x}{data frame to subset.}

\item{y}{reference data frame.}

\item{by}{names of columns to match.}
}
\value{
Data frame \code{x} with a subset of rows that have a \code{by}-wise
match in \code{y}. Columns of \code{x} are not modified.
}
\description{
Subset one data frame to match what is found in another.
}
