% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sexpf}
\alias{sexpf}
\title{Evaluate string with variables expanded.}
\usage{
sexpf(..., sep = " ", envir = parent.frame())
}
\arguments{
\item{...}{strings to expand variables in.}

\item{sep}{separator to join \code{...} by.}

\item{envir}{environment within which to evaluate strings.}
}
\value{
A string.
}
\description{
Variable definitions are of form \code{\{VAR:FORMAT\}},
where \code{FORMAT} corresponds to \code{\link{sprintf}} formats
without the leading percent sign. If \code{:FORMAT} is missing,
string format (\code{s}) is assumed. See examples for details.
}
\examples{
# Format value.
x = pi
sexpf("{x:.6f}")
# Format specification not needed for strings,
# or anything that cleanly renders to a string.
x = "foo"
sexpf("{x}")
# Split code to multiple lines.
x = 1
y = 2
z = 3
sexpf("{x:d}",
      "{y:d}",
      "{z:d}",
      sep=", ")

# Use double curly braces to avoid expansion.
size = 12L
sexpf("body {{ font-size: {size}px; }}")
}
