% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sexpf0}
\alias{sexpf0}
\title{Evaluate string with variables expanded.}
\usage{
sexpf0(..., sep = "", envir = parent.frame())
}
\arguments{
\item{...}{strings to expand variables in.}

\item{sep}{separator to join \code{...} by.}

\item{envir}{environment within which to evaluate strings.}
}
\value{
A string.
}
\description{
See \code{\link{sexpf}} for documentation.
}
