% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{slots}
\alias{slots}
\title{Values of slots.}
\usage{
slots(objs, name)
}
\arguments{
\item{objs}{a vector of objects with slots.}

\item{name}{a string, name of a \code{\link{slot}}.}
}
\value{
A vector of the values of slots of all objects.
}
\description{
Values of slots.
}
