% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{source.glob}
\alias{source.glob}
\title{Read R code from files.}
\usage{
source.glob(pattern = "*.R")
}
\arguments{
\item{pattern}{a glob pattern of filenames.}
}
\description{
Read and \code{\link{source}} all files that match \code{pattern}.
}
\seealso{
\code{\link{Sys.glob}} for glob patterns.
}
