% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{source.message}
\alias{source.message}
\title{Read R code from file.}
\usage{
source.message(file, ...)
}
\arguments{
\item{file}{name of file to read.}

\item{...}{passed to \code{\link{source}}.}
}
\description{
Read R code from file.
}
