% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{sp.to.points}
\alias{sp.to.points}
\title{Convert \pkg{sp} points to its data frame equivalent.}
\usage{
sp.to.points(shapes, fun = identity)
}
\arguments{
\item{shapes}{a \code{SpatialPointsDataFrame} object.}

\item{fun}{a function to apply to coordinates, e.g. \code{\link{round}}.}
}
\value{
A data frame.
}
\description{
Convert \pkg{sp} points to its data frame equivalent.
}
