% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{sp.to.polys}
\alias{sp.to.polys}
\title{Convert \pkg{sp} polygons to its data frame equivalent.}
\usage{
sp.to.polys(shapes, fun = identity, area.min = NULL,
  data.only = FALSE)
}
\arguments{
\item{shapes}{a \code{SpatialPolygonsDataFrame} object.}

\item{fun}{a function to apply to coordinates, e.g. \code{\link{round}}.}

\item{area.min}{threshold, polygons below which to discard.}

\item{data.only}{if \code{TRUE}, don't parse geometry, return only data.}
}
\value{
A list with two data frames: data and outlines.
}
\description{
Note that holes are marked in output, but not connected! To plot areas,
feed resulting polygons to \code{\link{attach.holes}}.
}
