% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{spatial.mean}
\alias{spatial.mean}
\title{2d spatially weighted mean.}
\usage{
spatial.mean(points, value, radius, weights = 1, na.rm = FALSE)
}
\arguments{
\item{points}{a data frame of points.}

\item{value}{a numeric vector of values.}

\item{radius}{radius of circle within which to weigh values.}

\item{weights}{a numeric vector of weights.}

\item{na.rm}{\code{TRUE} to ignore \code{NA} elements.}
}
\value{
A numeric vector of weighted values.
}
\description{
Values are weighted linearly with the weight at zero distance being one and
at radius distance being zero. These distance weights are further multiplied
by \code{weights} if given.
}
