% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{squeeze.polys}
\alias{squeeze.polys}
\title{Squeeze polygons to bounding box.}
\usage{
squeeze.polys(polys, xmin, xmax, ymin, ymax)
}
\arguments{
\item{polys}{a data frame of polygons.}

\item{xmin}{west edge of the bounding box.}

\item{xmax}{east edge of the bounding box.}

\item{ymin}{south edge of the bounding box.}

\item{ymax}{north edge of the bounding box.}
}
\value{
A data frame of polygons.
}
\description{
Squeeze polygons to bounding box.
}
