% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.mode}
\alias{stat.mode}
\title{Statistical mode.}
\usage{
stat.mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric object.}

\item{na.rm}{\code{TRUE} to ignore \code{NA} elements.}
}
\value{
A numeric vector of length one.
}
\description{
Statistical mode.
}
