% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stopif}
\alias{stopif}
\title{Ensure the truth of R expressions.}
\usage{
stopif(...)
}
\arguments{
\item{...}{expressions that should evaluate to \code{FALSE}.}
}
\description{
Ensure the truth of R expressions.
}
\seealso{
\code{\link{stopifnot}}
}
