% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{strafica.palette.diverging}
\alias{strafica.palette.diverging}
\title{Diverging palettes from Color Brewer.}
\description{
Diverging palettes from Color Brewer.
}
\references{
\url{http://colorbrewer2.org/}
}
