% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{strafica.palette.qualitative}
\alias{strafica.palette.qualitative}
\title{Qualitative palettes from Color Brewer.}
\description{
Qualitative palettes from Color Brewer.
}
\references{
\url{http://colorbrewer2.org/}
}
