% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{strafica.palette.sequential.multihue}
\alias{strafica.palette.sequential.multihue}
\title{Sequential multihue palettes from Color Brewer.}
\description{
Sequential multihue palettes from Color Brewer.
}
\references{
\url{http://colorbrewer2.org/}
}
