% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{strafica.palette.single.hue}
\alias{strafica.palette.single.hue}
\title{Sequential single hue palettes from Color Brewer.}
\description{
Sequential single hue palettes from Color Brewer.
}
\references{
\url{http://colorbrewer2.org/}
}
