% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{strwrap.str}
\alias{strwrap.str}
\title{Wrap strings with newlines.}
\usage{
strwrap.str(x, width, ...)
}
\arguments{
\item{x}{a character vector.}

\item{width}{width to wrap to.}

\item{...}{passed to \code{\link{strwrap}}.}
}
\value{
A character vector.
}
\description{
Wrap strings with newlines.
}
