% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sum.incomplete}
\alias{sum.incomplete}
\title{Join two data frames and sum values of one column.}
\usage{
\method{sum.incomplete}{not.an.s3.method}(df0, df1, by, name,
  missing = 0)
}
\arguments{
\item{df0}{data frame with first value for \code{name}.}

\item{df1}{data frame with second value for \code{name}.}

\item{by}{names of columns to join \code{df0} and \code{df1} by.}

\item{name}{name of column in \code{df0} and \code{df1} to sum.}

\item{missing}{value to use observations missing from one data frame.}
}
\value{
A data frame with columns \code{by} and \code{name}.
}
\description{
Join two data frames and sum values of one column.
}
