% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{system.message}
\alias{system.message}
\title{Invoke an OS command.}
\usage{
system.message(command, ...)
}
\arguments{
\item{command}{the system command to invoke.}

\item{...}{passed to \code{\link{system}}.}
}
\description{
Invoke an OS command.
}
