% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{tables.mdb}
\alias{tables.mdb}
\title{Names of tables in MS Access database.}
\usage{
tables.mdb(fname)
}
\arguments{
\item{fname}{name of database file.}
}
\value{
A character vector.
}
\description{
Uses ODBC, will fail on non-Windows systems.
}
