% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tb}
\alias{tb}
\title{Print the call stack of the last uncaught error.}
\usage{
tb(...)
}
\arguments{
\item{...}{passed to \code{\link{traceback}}.}
}
\description{
\code{tb} is an alias for \code{\link{traceback}}.
}
