% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map}
\alias{theme_map}
\title{A \pkg{ggplot2} theme for maps.}
\usage{
theme_map(base_size = 8, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{...}{passed to \code{\link{theme_strafica}}.}
}
\description{
A \pkg{ggplot2} theme for maps.
}
