% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_plain}
\alias{theme_plain}
\title{A simple \pkg{ggplot2} theme.}
\usage{
theme_plain(base_size = 8, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{...}{passed to \code{\link{theme_strafica}}.}
}
\description{
A simple \pkg{ggplot2} theme.
}
