% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_strafica}
\alias{theme_strafica}
\title{A generic \pkg{ggplot2} theme.}
\usage{
theme_strafica(base_size = 8, base_family = "", axis.x = TRUE,
  axis.y = TRUE, grid.colour = "grey83", grid.x = FALSE,
  grid.y = TRUE, grid.minor.x = FALSE, grid.minor.y = FALSE,
  legend.title = FALSE, line.colour = "grey25", panel.border = TRUE,
  panel.spacing = 0, strip.background = TRUE,
  strip.colour = "grey87")
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{axis.x}{\code{TRUE} to show the X-axis.}

\item{axis.y}{\code{TRUE} to show the Y-axis.}

\item{grid.colour}{colour of grid lines.}

\item{grid.x}{\code{TRUE} to show major vertical grid lines.}

\item{grid.y}{\code{TRUE} to show major horizontal grid lines.}

\item{grid.minor.x}{\code{TRUE} to show minor vertical grid lines.}

\item{grid.minor.y}{\code{TRUE} to show minor horizontal grid lines.}

\item{legend.title}{\code{TRUE} to show legend title.}

\item{line.colour}{colour of axis etc. lines.}

\item{panel.border}{\code{TRUE} to show panel border.}

\item{panel.spacing}{amount of space between panels.}

\item{strip.background}{\code{TRUE} to show a box behind panel titles.}

\item{strip.colour}{colour of panel title background.}
}
\description{
A generic \pkg{ggplot2} theme.
}
