% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{trace.grid}
\alias{trace.grid}
\title{Trace grid boundary to polygon outlines and areas.}
\usage{
trace.grid(grid, cell.size, by = NULL)
}
\arguments{
\item{grid}{a grid data frame.}

\item{cell.size}{size of grid cells.}

\item{by}{optional columns to group \code{grid} by.}
}
\value{
A list with three data frames: data, outlines and areas.
}
\description{
Trace grid boundary to polygon outlines and areas.
}
