% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{uind}
\alias{uind}
\title{Find indices of unique values.}
\usage{
uind(x)
}
\arguments{
\item{x}{a vector of values.}
}
\value{
A list of indices in \code{x}.
}
\description{
\code{uind} is much like \code{split(seq_along(x), x)},
but without using factors and thus also having the return
value in increasing order of index rather than factor level.
}
