% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unique.non.na}
\alias{unique.non.na}
\title{Discard \code{NA} and duplicate elements.}
\usage{
\method{unique}{non.na}(x, ...)
}
\arguments{
\item{x}{a vector.}

\item{...}{passed to \code{\link{unique}}.}
}
\value{
\code{x} with \code{NA} and duplicate elements discarded.
}
\description{
Discard \code{NA} and duplicate elements.
}
