% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{uniquify.lines}
\alias{uniquify.lines}
\title{Discard duplicated lines.}
\usage{
uniquify.lines(lines)
}
\arguments{
\item{lines}{a data frame of lines.}
}
\value{
A subset of \code{lines}.
}
\description{
A common use for \code{uniquify.lines} is with area boundaries, where the
boundary between two areas exists in two shapes and is often undesirable to
be plotted twice as it ruins linetype and alpha use.
}
