% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unpick}
\alias{unpick}
\title{Pick a column-wise subset of a data frame.}
\usage{
unpick(x, ...)
}
\arguments{
\item{x}{a data frame.}

\item{...}{names of columns to discard as either symbols or strings.}
}
\value{
A data frame.
}
\description{
Pick a column-wise subset of a data frame.
}
\seealso{
\code{\link{pick}}
}
