% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{view.lines}
\alias{view.lines}
\title{View lines on an interactive web map.}
\usage{
view.lines(lines, data = NULL, epsg = 4326)
}
\arguments{
\item{lines}{a data frame of lines.}

\item{data}{a data frame of per-entry attributes.}

\item{epsg}{projection code of input data, see
\code{\link{list.projections}}.}
}
\value{
A map object. (Usually you don't need this,
just call the function without assignment.)
}
\description{
View lines on an interactive web map.
}
