% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{view.points.mapbox}
\alias{view.points.mapbox}
\title{View points on an interactive web map.}
\usage{
view.points.mapbox(points, token, epsg = 4326)
}
\arguments{
\item{points}{a data frame of points.}

\item{token}{Mapbox access token.}

\item{epsg}{projection code of input data, see
\code{\link{list.projections}}.}
}
\value{
A map object. (Usually you don't need this,
just call the function without assignment.)
}
\description{
View points on an interactive web map.
}
