% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{view.polys.mapbox}
\alias{view.polys.mapbox}
\title{View polygons on an interactive web map.}
\usage{
view.polys.mapbox(polys, data = NULL, token, epsg = 4326)
}
\arguments{
\item{polys}{a data frame of polygons.}

\item{data}{a data frame of per-entry attributes.}

\item{token}{Mapbox access token.}

\item{epsg}{projection code of input data, see
\code{\link{list.projections}}.}
}
\value{
A map object. (Usually you don't need this,
just call the function without assignment.)
}
\description{
View polygons on an interactive web map.
}
