% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{view.segments}
\alias{view.segments}
\title{View segments on an interactive web map.}
\usage{
view.segments(segments, epsg = 4326)
}
\arguments{
\item{segments}{a data frame of segments.}

\item{epsg}{projection code of input data, see
\code{\link{list.projections}}.}
}
\value{
A map object. (Usually you don't need this,
just call the function without assignment.)
}
\description{
View segments on an interactive web map.
}
