% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysimp.R
\name{visvalingam}
\alias{visvalingam}
\title{Simplify polyline with the Visvalingam algorithm.}
\usage{
visvalingam(coords, epsilon, static = NULL, max.length = Inf)
}
\arguments{
\item{coords}{a two-column matrix of coordinates of polyline nodes.}

\item{epsilon}{threshold, below which nodes will be removed. While
Visvalingam's algorithm works on areas, \code{epsilon} here should be a
length, of similar magnitude as one would use with Douglas-Peucker.}

\item{static}{a logical vector indicating nodes which should not be removed.}

\item{max.length}{maximum length of line segments to form.}
}
\value{
A logical vector of rows to keep.
}
\description{
Simplify polyline with the Visvalingam algorithm.
}
\references{
\url{http://hydra.hull.ac.uk/resources/hull:8338},
\url{http://bost.ocks.org/mike/simplify/}
}
