% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{voronoi.area}
\alias{voronoi.area}
\title{Find discrete Voronoi regions.}
\usage{
voronoi.area(cx, cy, px, py, pid)
}
\arguments{
\item{cx}{X-coordinates of area centroids.}

\item{cy}{Y-coordinates of area centroids.}

\item{px}{X-coordinates of points.}

\item{py}{Y-coordinates of points.}

\item{pid}{point IDs.}
}
\value{
A vector of point IDs for each area.
}
\description{
For each area with centroids \code{cx}, \code{cy} find out which
of the points \code{px}, \code{py} is closest to the centroid.
}
