% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsubset.R
\name{vsubset}
\alias{vsubset}
\title{Return subset of data frames which meet conditions and reported number of rows dropped.
Wrapper function around \code{\link[base]{subset}}.}
\usage{
vsubset(df, ...)
}
\arguments{
\item{df}{data frame to be subsetted.}

\item{...}{passed on to \code{\link[base]{subset}}.}
}
\value{
A data frame.
}
\description{
Return subset of data frames which meet conditions and reported number of rows dropped.
Wrapper function around \code{\link[base]{subset}}.
}
