% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmap.R
\name{webmap.xdeg2num}
\alias{webmap.xdeg2num}
\title{Convert webmap WGS 84 X-coordinate to X-tile number.}
\usage{
webmap.xdeg2num(x, zoom)
}
\arguments{
\item{x}{WGS 84 X-coordinate.}

\item{zoom}{zoom level.}
}
\value{
X-tile number as integer.
}
\description{
Convert webmap WGS 84 X-coordinate to X-tile number.
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}
}
