% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmap.R
\name{webmap.xnum2deg}
\alias{webmap.xnum2deg}
\title{Convert webmap X-tile number to WGS 84 X-coordinate.}
\usage{
webmap.xnum2deg(xtile, zoom)
}
\arguments{
\item{xtile}{X-tile number.}

\item{zoom}{zoom level.}
}
\value{
WGS 84 X-coordinate as numeric.
}
\description{
Convert webmap X-tile number to WGS 84 X-coordinate.
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}
}
