% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmap.R
\name{webmap.ydeg2num}
\alias{webmap.ydeg2num}
\title{Convert webmap WGS 84 Y-coordinate to Y-tile number.}
\usage{
webmap.ydeg2num(y, zoom)
}
\arguments{
\item{y}{WGS 84 Y-coordinate.}

\item{zoom}{zoom level.}
}
\value{
Y-tile number as integer.
}
\description{
Convert webmap WGS 84 Y-coordinate to Y-tile number.
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames}
}
