% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{weighted.iqr}
\alias{weighted.iqr}
\title{Weighted interquartile range.}
\usage{
weighted.iqr(x, weights, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric object.}

\item{weights}{a numeric object of non-negative values.}

\item{na.rm}{\code{TRUE} to ignore \code{NA} elements.}
}
\value{
A numeric vector of length one.
}
\description{
Weighted interquartile range.
}
