% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{weighted.quantile}
\alias{weighted.quantile}
\title{Weighted quantiles.}
\usage{
weighted.quantile(x, weights, probs = seq(0, 1, 1/4), na.rm = FALSE)
}
\arguments{
\item{x}{a numeric object.}

\item{weights}{a numeric object of non-negative values.}

\item{probs}{a numeric vector of probabilities.}

\item{na.rm}{\code{TRUE} to ignore \code{NA} elements.}
}
\value{
A numeric vector same length as \code{probs}.
}
\description{
\code{weighted.quantile} does piecewise linear interpolation between
midpoints of steps of the empirical cdf. This matches the behaviour of
\code{\link{quantile}} given argument \code{type=5}.
}
