% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{Print summary of data frame contents.}
\usage{
what(df, ...)
}
\arguments{
\item{df}{a data frame or a vector.}

\item{...}{names of columns as either symbols or strings.
If not given, summaries of all columns will be printed.}
}
\description{
Print summary of data frame contents.
}
\seealso{
\code{\link{summary}}, \code{\link{str}}
}
