% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.delim.t}
\alias{write.delim.t}
\title{Write transpose to tab-delimited file.}
\usage{
write.delim.t(x, fname, quote = FALSE, sep = "\\t", na = "",
  row.names = TRUE, col.names = FALSE, ...)
}
\arguments{
\item{x}{The object to be written, preferably a matrix or data frame.}

\item{fname}{A character string naming a file.}

\item{quote}{A logical value (\code{TRUE} or \code{FALSE}) or a numeric 
vector. If \code{TRUE}, any character or factor columns will be surrounded 
by double quotes. If a numeric vector, its elements are taken as the
indices of columns to quote. In both cases, row and column names are quoted
if they are written. If \code{FALSE}, nothing is quoted.}

\item{sep}{The field separator string.}

\item{na}{The string to use for missing values in the data.}

\item{row.names}{Either a logical value indicating whether the row names of
\code{x} are to be written along with \code{x}, or a character vector of
row names to be written.}

\item{col.names}{Either a logical value indicating whether the column names 
of \code{x} are to be written along with \code{x}, or a character vector of
column names to be written.}

\item{...}{Further arguments to be passed to
\code{\link[utils]{write.table}}.}
}
\description{
See \code{\link{write.table}} for arguments.
}
