% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{write.emme.matrices}
\alias{write.emme.matrices}
\title{Write matrices to Emme matrix file.}
\usage{
write.emme.matrices(p, q, value, fname, header = "")
}
\arguments{
\item{p}{a list of vectors of row values (e.g. origins).}

\item{q}{a list of vectors of column values (e.g. destinations).}

\item{value}{a list of vectors of values for p-q pairs.}

\item{fname}{name of file to write.}

\item{header}{a vector of values of matrix headers.
Do not include "t matrices" here, it is added automatically.}
}
\description{
Write matrices to Emme matrix file.
}
