% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emme.R
\name{write.emme.matrix}
\alias{write.emme.matrix}
\title{Write matrix to Emme matrix file.}
\usage{
write.emme.matrix(p, q, value, fname, header = "")
}
\arguments{
\item{p}{a vector of row values (e.g. origins).}

\item{q}{a vector of column values (e.g. destinations).}

\item{value}{a vector of values for p-q pairs.}

\item{fname}{name of file to write.}

\item{header}{a matrix header string.
Do not include "t matrices" here, it is added automatically.}
}
\description{
Write matrix to Emme matrix file.
}
