% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{write.geojson}
\alias{write.geojson}
\title{This function has been removed from the package. Use `sf::st_write()`
instead.}
\usage{
write.geojson(shapes, fname, epsg)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of GeoJSON file to write.}

\item{epsg}{projection code of input data, see
\code{\link{list.projections}}. Output is always WGS 84 and
automatically reprojected to it if needed.}
}
\description{
This function has been removed from the package. Use `sf::st_write()`
instead.
}
