% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg.R
\name{write.gpkg}
\alias{write.gpkg}
\title{This function has been removed from the package. Use `sf::st_write()`
instead.}
\usage{
write.gpkg(shapes, fname, epsg = NULL)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of GeoPackage file to write.}

\item{epsg}{projection code, see \code{\link{list.projections}}.}
}
\description{
This function has been removed from the package. Use `sf::st_write()`
instead.
}
