% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{read.mapinfo}
\alias{read.mapinfo}
\title{Read points, lines or polygons from a MapInfo TAB file.}
\usage{
read.mapinfo(fname, stringsAsFactors = FALSE, encoding = "CP1252", ...)
}
\arguments{
\item{fname}{name of MapInfo TAB file.}

\item{stringsAsFactors}{should character vector be converted to factors?}

\item{encoding}{character encoding used in code{fname}.}

\item{...}{passed to \code{\link[rgdal]{readOGR}}.}
}
\value{
A \code{Spatial*DataFrame} object.
}
\description{
Read points, lines or polygons from a MapInfo TAB file.
}
