% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkg.R
\name{write.gpkg}
\alias{write.gpkg}
\title{Write points, lines or polygons to a GeoPackage file.}
\usage{
write.gpkg(shapes, fname, epsg = NULL)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of GeoPackage file to write.}

\item{epsg}{projection code, see \code{\link{list.projections}}.}
}
\description{
Write points, lines or polygons to a GeoPackage file.
}
