% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{write.mapinfo}
\alias{write.mapinfo}
\title{Write points, lines or polygons to a MapInfo TAB file.}
\usage{
write.mapinfo(shapes, fname, epsg)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of MapInfo TAB file to write.}

\item{epsg}{projection code, see \code{\link{list.projections}}.}
}
\description{
Requires \code{ogr2ogr} in \code{$PATH}.
}
