% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapinfo.R
\name{write.mif}
\alias{write.mif}
\title{This function has been removed from the package. Use `sf::st_write()`
instead.}
\usage{
write.mif(shapes, fname, epsg)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of the MapInfo MIF file to write.}

\item{epsg}{projection code, see \code{\link{list.projections}}.}
}
\description{
This function has been removed from the package. Use `sf::st_write()`
instead.
}
