% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{write.shape}
\alias{write.shape}
\title{This function has been removed from the package. Use `sf::st_write()`
instead.}
\usage{
write.shape(shapes, fname, epsg, verbose = TRUE)
}
\arguments{
\item{shapes}{a \code{Spatial*DataFrame} object.}

\item{fname}{name of shapefile to write.}

\item{epsg}{projection code, see \code{\link{list.projections}}.}

\item{verbose}{a logical indicating whether or not to print out progress updates.}
}
\description{
This function has been removed from the package. Use `sf::st_write()`
instead.
}
